backend/requirements.txt
# RAG Node Dependencies

# Core LangChain - Install main package first
langchain>=0.3.0

# LangChain Components
langchain-community>=0.3.0
langchain-core>=0.3.0
langchain-text-splitters>=0.3.0
langchain-openai>=0.2.0
langchain-groq>=0.2.0

# Vector Store
faiss-cpu>=1.8.0
# For GPU support, use: faiss-gpu>=1.8.0

# Embeddings & ML
sentence-transformers>=2.2.2
transformers>=4.30.0
torch>=2.0.0
numpy>=1.24.0
huggingface-hub>=0.20.0
langchain-huggingface>=0.1.0

# LLM Providers
groq>=0.11.0
openai>=1.0.0

# Document Loading
pypdf>=3.17.0
pypdf2>=3.0.0

# Additional Dependencies
pydantic>=2.0.0
pydantic-settings>=2.0.0
typing-extensions>=4.7.0
python-dotenv>=1.0.0
tiktoken>=0.5.2

# HTTP and async
httpx>=0.24.0
aiohttp>=3.9.0
requests>=2.31.0

# Data processing
pandas>=2.0.0

# FastAPI & Web Server
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# LangGraph
langgraph>=0.2.0
langgraph-checkpoint>=1.0.0

# LiveKit Voice Pipeline (v1.x)
livekit>=1.0.20
livekit-agents>=1.3.6
livekit-plugins-silero>=1.3.6
livekit-plugins-groq>=1.3.6
